refactor / clean-up
