+++ src/webWorkers/proteinFile/index.worker.js
@@ -1,11 +1,15 @@
+// @flow
 import 'babel-polyfill';
 import fetch from 'isomorphic-fetch';
 import { stringify as qsStringify } from 'query-string';
 
 import description2path from 'utils/processLocation/description2path';
 
+// Max page size provided by the server
+// to maximise the number of results sent by the server at once
 const MAX_PAGE_SIZE = 200;
 
+// From a pathname and search parameter, generates a full URL
 const getUrl = (pathname, taxId, page) =>
   `${pathname}?${qsStringify({
     tax_id: taxId,
@@ -13,10 +17,12 @@ const getUrl = (pathname, taxId, page) =>
     page_size: MAX_PAGE_SIZE,
   })}`;
 
+// Helper function to send progress information back to the main thread
 const progress = (value /*: number */) => {
   self.postMessage({ type: 'progress', details: value });
 };
 
+// eslint-disable-next-line max-statements
 const processEvent = async ({ data: { description, api, taxId, type } }) => {
   const content = [];
   const pathname = `${api.protocol}//${api.hostname}:${api.port}${api.root}${description2path(
@@ -36,13 +42,12 @@ const processEvent = async ({ data: { description, api, taxId, type } }) => {
   let totalCount;
   let next = getUrl(pathname, taxId, page);
   while (next) {
-    console.log(`getting ${next}`);
     const response = await fetch(next);
     const obj = await response.json();
     totalCount = obj.count + 1;
     next = obj.next && getUrl(pathname, taxId, ++page);
     for (const {
-      metadata: { accession, source_database: db, name: { name } },
+      metadata: { accession, source_database: db, name },
     } of obj.results) {
       if (type === 'accession') {
         content.push(accession);
@@ -67,18 +72,20 @@ const processEvent = async ({ data: { description, api, taxId, type } }) => {
   return url;
 };
 
-let activated = false;
-
-self.addEventListener('message', async e => {
-  if (activated) return;
-  activated = true;
-  let url;
-  try {
-    url = await processEvent(e);
-  } catch (error) {
-    self.postMessage({ type: 'failed', details: error });
-  }
-  self.postMessage({ type: 'success', details: url });
-  // Don't close it now, otherwise the blob will disappear
-  // Let it be terminated by the main thread on component unmount
-});
+self.addEventListener(
+  'message',
+  async e => {
+    let url;
+    try {
+      url = await processEvent(e);
+    } catch (error) {
+      self.postMessage({ type: 'failed', details: error });
+    }
+    self.postMessage({ type: 'success', details: url });
+    // Don't close it now, otherwise the blob will disappear
+    // Let it be terminated by the main thread on component unmount
+    // NOTE: should maybe transfer the blob back to main thread to be able to kill
+    // the worker without losing the file
+  },
+  { once: true },
+);
